some stuff fixed and new stuff

+++ src/handlers/dataBaseCleanUp.js
@@ -16,6 +16,21 @@ cron.schedule('0 0 * * *', async () => {
 });
 
 async function cleanupExpiredData(now) {
-  const tables = ['GuildSettings']
-  await query(`DELETE FROM ${table} WHERE deletionDate IS NOT NULL AND deletionDate < ?`, [now])
+  const tables = ['GuildSettings', 'LevelSystem', 'LevelSettings', 'EventEmbeds', 'GeneratedEmbeds', 'ReactionRoles'];
+  const cleanupInfo = [];
+
+  for (const table of tables) {
+    const [rowsToDelete] = await query(`SELECT COUNT(*) AS count FROM ${table} WHERE deletionDate IS NOT NULL AND deletionDate < ?`, [now]);
+    const rowsDeleted = rowsToDelete[0]?.count || 0;
+
+    if (rowsDeleted > 0) {
+      await query(`DELETE FROM ${table} WHERE deletionDate IS NOT NULL AND deletionDate < ?`, [now]);
+    }
+
+    cleanupInfo.push({
+      table: table,
+      rowsDeleted: rowsDeleted
+    });
+  }
+  return cleanupInfo;
 }
